+++ koolreport/datasources/SQLSRVDataSource.php
@@ -25,7 +25,8 @@ use \koolreport\core\Utility;
 class SQLSRVDataSource extends DataSource
 {
 	protected $connection;
-	protected $query;
+    protected $query;
+    protected $sqlParams;
 	protected function onInit()
 	{		
     $host = Utility::get($this->params,"host","");//host\instanceName
@@ -43,27 +44,27 @@ class SQLSRVDataSource extends DataSource
     }
 	}
 	
-	public function query($query, $params=null)
+	public function query($query, $sqlParams=null)
 	{
 		$this->query = $query;
-    if($params!=null)
+        if($sqlParams!=null)
 		{
-			$this->params = $params;
+			$this->sqlParams = $sqlParams;
 		}
 		return $this;
 	}
   
-  public function params($params)
+    public function params($sqlParams)
 	{
-		$this->params = $params;
+		$this->sqlParams = $sqlParams;
 		return $this;
 	}
 	
-	protected function bindParams($query,$params)
+	protected function bindParams($query,$sqlParams)
 	{
-		if($params!=null)
+		if($sqlParams!=null)
 		{
-			foreach($params as $key=>$value)
+			foreach($sqlParams as $key=>$value)
 			{
 				if(gettype($value)==="array")
 				{
@@ -113,7 +114,7 @@ class SQLSRVDataSource extends DataSource
 	
 	public function start()
 	{
-    $query = $this->bindParams($this->query,$this->params);
+    $query = $this->bindParams($this->query,$this->sqlParams);
     $stmt = sqlsrv_query( $this->connection, $query);
     if( $stmt === false ) 
          die( print_r( sqlsrv_errors(), true));