+++ koolreport/CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 1.61.5
+    1. Fix parameters bug in PDODataSource
+    1. Fix parameters bug in MySQLDataSource
+    1. Fix parameters bug in SQLSRVDataSource
+
 # Version 1.61.2
     1. Add html() method to Widget to allow return html of widget
     2. Add $return to create() method of Widget to return html instead of echo html.